fix: resolve 8 critical/high priority review findings (#5,#6,#7,#8,#18,#20,#21,#22)
- #5: Add bounds check in hex parser to prevent buffer overflow
- #6: Fix ISR race conditions in S0 pulse counter (volatile, critical section)
- #7: HTML-escape URI in sendApiNotFound() to prevent reflected XSS
- #8: Remove debug gate from evalOutputs() so GPIO outputs work normally
- #18: Add null pointer checks on strtok() results in MQTT callback
- #20: Move file open after existence check to prevent file descriptor leak
- #21: Change year variable from int8_t to int16_t to prevent overflow
- #22: Use async Dallas temperature reading to avoid 750ms blocking

+++ src/OTGW-firmware/helperStuff.ino
@@ -411,7 +411,7 @@ bool yearChanged(){
   static int16_t lastyear = -1;
   TimeZone myTz =  timezoneManager.createForZoneName(CSTR(settingNTPtimezone));
   ZonedDateTime myTime = ZonedDateTime::forUnixSeconds64(time(nullptr), myTz);
-  int8_t thisyear = myTime.year();
+  int16_t thisyear = myTime.year();
   bool _ret = (lastyear != thisyear); //year changed
   if (_ret) {
     //year changed
